# weft/src/api/openapi.yaml
openapi: 3.0.0
info:
  title: Weft API
  version: 0.2.0
  description: |
    Coordinator service for Loominal multi-agent infrastructure.

    Weft routes work, manages agent lifecycle, handles scaling, and provides a REST API
    for fleet management. It runs as a multi-tenant service that handles multiple projects
    in a single deployment.

    ## Key Features
    - **Agent Management**: Register, discover, and shutdown agents
    - **Work Distribution**: Submit and track work items across agents
    - **Spin-Up Targets**: Manage dynamic agent provisioning
    - **Batch Operations**: Perform bulk operations on agents, targets, and work
    - **Pagination**: All list endpoints support cursor-based pagination
    - **Semantic Context**: Responses include resolved agent details for better decision-making

  contact:
    name: Michael LoPresti
    email: mike@lopresti.org

  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://weft.loominal.dev
    description: Production server
...
paths:
...
  /api/work:
    get:
      summary: List work items
      description: |
        Returns a paginated list of work items with optional filtering.
        Includes semantic context with resolved agent details when work is assigned.
      operationId: listWork
      tags:
        - Work
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          description: Filter by work status
          schema:
            $ref: '#/components/schemas/WorkItemStatus'
        - name: boundary
          in: query
          description: Filter by boundary/classification
          schema:
            type: string
            example: personal
        - name: classification
          in: query
          deprecated: true
          description: (Deprecated) Use 'boundary' instead
          schema:
            type: string
        - name: cursor
          in: query
          description: Pagination cursor from previous response
          schema:
            type: string
        - name: limit
          in: query
          description: Maximum number of items to return
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        '200':
          description: List of work items with pagination
          headers:
            X-Deprecated-Param:
              description: Indicates use of deprecated parameter
              schema:
                type: string
                example: classification (use boundary instead)
          content:
            application/json:
              schema:
                type: object
                required:
                  - workItems
                  - count
                  - hasMore
                properties:
                  workItems:
                    type: array
                    items:
                      $ref: '#/components/schemas/WorkItemResponse'
                  count:
                    type: integer
                    example: 50
                  total:
                    type: integer
                    example: 200
                  hasMore:
                    type: boolean
                    example: true
                  nextCursor:
                    type: string
                    nullable: true
                  prevCursor:
                    type: string
                    nullable: true
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      summary: Submit new work item
      description: |
        Submits a new work item to the coordinator for routing and assignment.
        The coordinator will match the work to available agents or trigger spin-up.
      operationId: submitWork
      tags:
        - Work
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkSubmitRequest'
            example:
              taskId: task-123
              boundary: personal
              capability: typescript
              description: Implement feature X
              priority: 7
              preferredAgentType: claude-code
      responses:
        '201':
          description: Work item created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkSubmitResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/work/{id}:
    get:
      summary: Get work item details
      description: |
        Returns detailed information about a specific work item including
        semantic context with resolved agent details if assigned.
      operationId: getWorkItem
      tags:
        - Work
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Work item ID
          schema:
            type: string
            example: abc-123
      responses:
        '200':
          description: Work item details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkItemResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/work/{id}/cancel:
    post:
      summary: Cancel work item
      description: Cancels a specific work item if it hasn't been completed
      operationId: cancelWorkItem
      tags:
        - Work
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Work item ID to cancel
          schema:
            type: string
      responses:
        '200':
          description: Work item cancelled successfully
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - message
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Work item abc-123 cancelled
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/work/cancel-batch:
    post:
      summary: Batch cancel work items
      description: |
        Cancels multiple work items matching filter criteria or specific IDs.
        Optionally reassigns cancelled work to other agents.
      operationId: batchCancelWork
      tags:
        - Work
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchCancelWorkRequest'
            examples:
              filterBased:
                summary: Cancel pending low-priority work
                value:
                  filter:
                    status: pending
                    minPriority: 3
                  reason: resource-constraint
              guidBased:
                summary: Cancel specific work items
                value:
                  workItemIds:
                    - task-1
                    - task-2
                  reason: user-requested
      responses:
        '200':
          description: Batch operation results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchCancelWorkResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Bearer token authentication (configure via API_TOKENS env var)
  schemas:
    AgentType:
      type: string
      enum:
        - copilot-cli
        - claude-code
      description: Type of AI agent
...
    WorkItemStatus:
      type: string
      enum:
        - pending
        - assigned
        - in-progress
        - completed
        - failed
        - cancelled
      description: Status of a work item
...
    Priority:
      type: integer
      minimum: 1
      maximum: 10
      description: Priority level (1=lowest, 10=highest)
    AgentSummary:
      type: object
      description: Summary information about an agent for semantic context
      required:
        - guid
        - agentType
      properties:
        guid:
          type: string
          format: uuid
          description: Unique agent identifier
        handle:
          type: string
          description: Agent handle/username
        agentType:
          $ref: '#/components/schemas/AgentType'
        hostname:
          type: string
          description: Hostname where agent is running
...
    WorkItemResponse:
      type: object
      description: Work item with semantic context (resolved agent details)
      required:
        - id
        - taskId
        - capability
        - description
        - priority
        - boundary
        - status
        - offeredBy
        - offeredAt
        - attempts
      properties:
        id:
          type: string
        taskId:
          type: string
        capability:
          type: string
        description:
          type: string
        priority:
          $ref: '#/components/schemas/Priority'
        boundary:
          type: string
          description: Workload boundary for routing
        preferredAgentType:
          $ref: '#/components/schemas/AgentType'
        requiredAgentType:
          $ref: '#/components/schemas/AgentType'
        status:
          $ref: '#/components/schemas/WorkItemStatus'
        assignedTo:
          type: string
          format: uuid
          description: Assigned agent GUID
        assignedToAgent:
          $ref: '#/components/schemas/AgentSummary'
          description: Resolved agent details if assigned
        assignedAt:
          type: string
          format: date-time
        progress:
          type: integer
          minimum: 0
          maximum: 100
        offeredBy:
          type: string
        offeredAt:
          type: string
          format: date-time
        attempts:
          type: integer
        deadline:
          type: string
          format: date-time
        contextData:
          type: object
          additionalProperties: true
    WorkSubmitRequest:
      type: object
      required:
        - taskId
        - boundary
        - capability
        - description
      properties:
        taskId:
          type: string
          description: Application-defined task identifier
        boundary:
          type: string
          description: Workload boundary for routing
        capability:
          type: string
          description: Required capability
        description:
          type: string
          description: Human-readable task description
        priority:
          $ref: '#/components/schemas/Priority'
          default: 5
        preferredAgentType:
          $ref: '#/components/schemas/AgentType'
        requiredAgentType:
          $ref: '#/components/schemas/AgentType'
        deadline:
          type: string
          format: date-time
        contextData:
          type: object
          additionalProperties: true
    WorkSubmitResponse:
      type: object
      required:
        - id
        - workItemId
        - targetAgentType
        - spinUpTriggered
      properties:
        id:
          type: string
          description: Generated work item ID
        workItemId:
          type: string
          deprecated: true
          description: Use 'id' instead
        targetAgentType:
          $ref: '#/components/schemas/AgentType'
        spinUpTriggered:
          type: boolean
        estimatedWaitSeconds:
          type: integer
...
    BatchCancelWorkRequest:
      type: object
      properties:
        filter:
          type: object
          properties:
            status:
              type: string
            boundary:
              type: string
            capability:
              type: string
            minPriority:
              type: integer
            assignedTo:
              type: string
              format: uuid
        workItemIds:
          type: array
          items:
            type: string
        reason:
          type: string
          enum: [user-requested, deadline-passed, resource-constraint, system-shutdown]
        reassign:
          type: boolean
          default: false
    BatchCancelWorkResponse:
      type: object
      required:
        - success
        - failed
        - count
        - errors
        - completedAt
        - totalProcessed
        - successRate
        - cancelledItems
        - reassignedItems
        - notCancellable
      properties:
        success:
          type: array
          items:
            type: string
        failed:
          type: array
          items:
            type: string
        count:
          type: integer
        errors:
          type: object
          additionalProperties:
            type: string
        completedAt:
          type: string
          format: date-time
        totalProcessed:
          type: integer
        successRate:
          type: number
        cancelledItems:
          type: array
          items:
            type: string
        reassignedItems:
          type: array
          items:
            type: string
        notCancellable:
          type: array
          items:
            type: string
...
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error type
        message:
          type: string
          description: Error message
        details:
          type: object
          description: Additional error details
  responses:
    BadRequest:
      description: Bad request - invalid parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: ValidationError
            message: Invalid request parameters
    Unauthorized:
      description: Unauthorized - missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Unauthorized
            message: Invalid or missing authentication token
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: NotFound
            message: Resource not found
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: InternalServerError
            message: An unexpected error occurred